Faithfully render braces in generated markup
By interpreting partials as Liquid templates, the `render_partial`
plugin strips content that happens to conform to that engine's syntax.
Specifically, the referenced documentation describes a substitution
syntax for Web Platform Test files which reserves the `{{` and `}}`
strings as delimiters for variable interpolation.

This plugin is only used to render content that is intended to stand
alone (i.e. without variable expansion), making the use of a templating
engine superfluous.

Remove the use of the rendering engine and emit the referenced text
directly.